use an explicit cast to int32_t for arguments to rand48 (avoids ambiguity for Metrowerks, which defines int32_t as long)
git-svn-id: http://svn.boost.org/svn/boost/trunk@7935 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/libs/random/random_test.cpp
@@ -174,9 +174,9 @@ void instantiate_all()
 
 #if !defined(BOOST_NO_INT64_T) && !defined(BOOST_NO_INTEGRAL_INT64_T)
   instantiate_urng("rand48", rand48(), 0);
-  rand48 rnd(5);
+  rand48 rnd(int32_t(5));
   rand48 rnd2(uint64_t(0x80000000) * 42);
-  rnd.seed(17);
+  rnd.seed(int32_t(17));
   rnd2.seed(uint64_t(0x80000000) * 49);
 #endif
